+++ libs/plant/view/src/lib/plant-page/plant-page.tsx
@@ -3,26 +3,7 @@ import React from 'react';
 import styled from '@emotion/styled';
 
 import { Plant } from '@gardn/plant/smart';
-
-import { withRouter } from 'react-router-dom'
-const StyledBackButton = styled.div`
-  background-color: green;
-
-  position: absolute;
-  left: .5rem;
-  top: .5rem;
-  height: 1rem;
-  width: 1rem;
-`;
-export const GoBack = withRouter(({ history }) => {
-  return (
-    <StyledBackButton>
-      <button onClick={() => history.goBack()}>
-        &#x3C;
-      </button>
-    </StyledBackButton>
-  )
-});
+import { GoBack } from '@gardn/utility';
 
 
 const StyledPlantPage = styled.div`
