+++ libs/utility/.babelrc
@@ -1,0 +1,4 @@
+{
+  "presets": ["@nrwl/react/babel", "@emotion/babel-preset-css-prop"],
+  "plugins": []
+}
