+++ libs/utility/README.md
@@ -1,0 +1,7 @@
+# utility
+
+This library was generated with [Nx](https://nx.dev).
+
+## Running unit tests
+
+Run `nx test utility` to execute the unit tests via [Jest](https://jestjs.io).
